Added code to wallet page instead home page
